backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
ccxt>=4.1.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
itsdangerous>=2.1.2
cryptography>=42.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
